# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервиса PTOP",
        "title": "PTOP API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/ws/offers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Подключение для получения событий CRUD по офферам.",
                "tags": [
                    "offers"
                ],
                "summary": "WebSocket обновления офферов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "канал",
                        "name": "channel",
                        "in": "query"
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "$ref": "#/definitions/handlers.OfferEvent"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ws/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "После подключения авторизованный клиент получает события OrderEvent о создании своих ордеров.",
                "tags": [
                    "orders"
                ],
                "summary": "Websocket ордеров клиента",
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "$ref": "#/definitions/handlers.OrderEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ws/orders/{id}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Позволяет автору и владельцу оффера получать события OrderStatusEvent при каждом изменении статуса указанного ордера.",
                "tags": [
                    "orders"
                ],
                "summary": "Websocket уведомлений о статусе ордера",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID ордера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "$ref": "#/definitions/handlers.OrderStatusEvent"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.OfferEvent": {
            "type": "object",
            "properties": {
                "offer": {
                    "$ref": "#/definitions/models.OfferFull"
                },
                "type": {
                    "type": "string",
                    "example": "created"
                }
            }
        },
...
        "handlers.OrderEvent": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/models.OrderFull"
                },
                "type": {
                    "type": "string",
                    "example": "order.created"
                }
            }
        },
...
        "handlers.OrderStatusEvent": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/models.OrderFull"
                },
                "type": {
                    "type": "string",
                    "example": "order.status_changed"
                }
            }
        },
...
        "models.Asset": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isConvertible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.Client": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "ordersCount": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "registredAt": {
                    "type": "string"
                },
                "twoFAEnabled": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.ClientPaymentMethod": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "clientID": {
                    "type": "string"
                },
                "country": {
                    "$ref": "#/definitions/models.Country"
                },
                "countryID": {
                    "type": "string"
                },
                "detailedInformation": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paymentMethod": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "paymentMethodID": {
                    "type": "string"
                },
                "postCode": {
                    "type": "string"
                }
            }
        },
        "models.Country": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.FeeSideType": {
            "type": "string",
            "enum": [
                "sender",
                "receiver",
                "shared"
            ],
            "x-enum-varnames": [
                "FeeSideSender",
                "FeeSideReceiver",
                "FeeSideShared"
            ]
        },
        "models.KycLevelHintType": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high"
            ],
            "x-enum-varnames": [
                "KycLevelHintLow",
                "KycLevelHintMedium",
                "KycLevelHintHigh"
            ]
        },
...
        "models.Offer": {
            "type": "object",
            "properties": {
                "TTL": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "clientID": {
                    "type": "string"
                },
                "conditions": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "disabledAt": {
                    "type": "string"
                },
                "enabledAt": {
                    "type": "string"
                },
                "fromAssetID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "maxAmount": {
                    "type": "number"
                },
                "minAmount": {
                    "type": "number"
                },
                "orderExpirationTimeout": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "toAssetID": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.OfferFull": {
            "type": "object",
            "properties": {
                "TTL": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "client": {
                    "$ref": "#/definitions/models.Client"
                },
                "clientID": {
                    "type": "string"
                },
                "clientPaymentMethods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ClientPaymentMethod"
                    }
                },
                "conditions": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "disabledAt": {
                    "type": "string"
                },
                "enabledAt": {
                    "type": "string"
                },
                "fromAsset": {
                    "$ref": "#/definitions/models.Asset"
                },
                "fromAssetID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "maxAmount": {
                    "type": "number"
                },
                "minAmount": {
                    "type": "number"
                },
                "orderExpirationTimeout": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "toAsset": {
                    "$ref": "#/definitions/models.Asset"
                },
                "toAssetID": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.OrderFull": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "author": {
                    "$ref": "#/definitions/models.Client"
                },
                "authorID": {
                    "type": "string"
                },
                "buyer": {
                    "$ref": "#/definitions/models.Client"
                },
                "buyerID": {
                    "type": "string"
                },
                "clientPaymentMethod": {
                    "$ref": "#/definitions/models.ClientPaymentMethod"
                },
                "clientPaymentMethodID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "fromAsset": {
                    "$ref": "#/definitions/models.Asset"
                },
                "fromAssetID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isEscrow": {
                    "type": "boolean"
                },
                "offer": {
                    "$ref": "#/definitions/models.Offer"
                },
                "offerID": {
                    "type": "string"
                },
                "offerOwner": {
                    "$ref": "#/definitions/models.Client"
                },
                "offerOwnerID": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "releasedAt": {
                    "type": "string"
                },
                "seller": {
                    "$ref": "#/definitions/models.Client"
                },
                "sellerID": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.OrderStatus"
                },
                "toAsset": {
                    "$ref": "#/definitions/models.Asset"
                },
                "toAssetID": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.OrderStatus": {
            "type": "string",
            "enum": [
                "WAIT_PAYMENT",
                "PAID",
                "RELEASED",
                "CANCELLED",
                "DISPUTE"
            ],
            "x-enum-varnames": [
                "OrderStatusWaitPayment",
                "OrderStatusPaid",
                "OrderStatusReleased",
                "OrderStatusCancelled",
                "OrderStatusDispute"
            ]
        },
        "models.PaymentMethod": {
            "type": "object",
            "properties": {
                "chargebackWindowHours": {
                    "type": "integer"
                },
                "countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Country"
                    }
                },
                "feeSide": {
                    "$ref": "#/definitions/models.FeeSideType"
                },
                "id": {
                    "type": "string"
                },
                "isRealtime": {
                    "type": "boolean"
                },
                "isReversible": {
                    "type": "boolean"
                },
                "kycLevelHint": {
                    "$ref": "#/definitions/models.KycLevelHintType"
                },
                "methodGroup": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "regions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "settlementMinutes": {
                    "type": "integer"
                },
                "typicalFiatCCY": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}